ai/requirements.txt
flask
google-generativeai
transformers
python-dotenv
Pillow
